I fixed the problem with the placement buttons not responding.  We needed the intent to be changed the to ship placement class when single player was selected.
Right now the most obvious error is that, when you click "Ok" to close the "Place your ships" dialog, it automatically clicks the image button behind it.

Another error I've noted is a nullPointerException whenever we reference a variable from another class using a class object.  For example, right now there is a nullPointerException on any placement button click because the program does not recognize:

main.player1ShipArray

This needs to be fixed soon.